requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pyjwt
pydantic
python-multipart
